backend/requirements.dev.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.0
Django==5.0.2
django-cors-headers==4.3.1
django-picklefield==3.1
django-protected-media==1.0.2
django-q2==1.6.1
djangorestframework==3.14.0
fastapi==0.109.2
firebase-admin==6.5.0
frozenlist==1.4.1
google-api-core==2.19.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.63.0
grpcio-status==1.62.2
httplib2==0.22.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
msgpack==1.0.0
multidict==6.0.5
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.2
pydantic_core==2.16.3
PyJWT==2.8.0
pyOpenSSL==24.0.0
pyparsing==3.1.2
pytz==2024.1
PyVimeo==1.1.2
requests==2.31.0
rsa==4.9
service-identity==24.1.0
setuptools==69.5.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.36.3
tinydb==4.8.0
tuspy==1.0.3
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
yarl==1.9.4
zope.interface==6.2.0

# Development packages
django-debug-toolbar
django-filter
drf-standardized-errors
pyyaml 
inflection